Removal of pixiv.net styles (read des for info)
The only style that was present was one that removed the page change links that covered the content. Upon clicking the links the page would scroll to another page or image. The removal of these links results in many other features of the site breaking such as the reader.

If anybody has a more concrete way of removing the links without breaking any feature of the site, I would be more than happy to re-implement it.