New versioning system (read desc of commit)
When implemenenting a new style for Stylus itself, I came accross an
issue which resulted in styles being uninstallable on Opera when using
my previous versioning system. Instead of versions following the rulset
of N.N it will now follow N.N.N.

N.N was an incremental versioning system with no real logic.
Once the second N got to a value of ten or higher the first N would
increment.

Examples of valid version numbers for the N.N system:
1.0
1.1
2.0
2.1

The version below would be invalid as instead it should increment the
first N to two and the second N should reset to zero.
1.10

N.N.N has a more logical incremental approach. In a way it is similar
to the N.N system except that the third N is for minor bug fixes to
styles. Once a style recieves a significant change/update, the second N
will increment. If a typo was rectified the Third N will increment. The
third N can increment way into the hundreds if needbe. Once the second
N increments the third N will reset to zero. The second N cannot go
over nine and if it was to, the first N will increment while also
resetting the second N to zero.

Examples of valid version numbers for the N.N.N system:
1.0.1
1.1.1
1.1.0
2.0.1
2.0.0
2.1.1
2.0.10
1.2.45

The version below would be invalid as instead it should increment the
first N to two and the second and third N should reset to zero.
1.10.45