new style for youtube.com
Honestly youtube please can we fucking not with this shit, why show me a video in my recommended if its only watchable by people who are members of the channel who uploaded it. If i was a member then sure, show away. but im not, and thats for a fucking reason. piss off with your shitty UI/UX design choices and just be a good fucking platform for content again. your policies result in youtubers having to tone down their content alot of the time and it defeats the point of people being THEMSELVES. not to mention your insanely exploitable copyright system, fuck off with the automation and manually review content, or fuck off with copyright strikes as a whole.